+++ src/client/commands/run.rs
@@ -4,8 +4,8 @@ use crate::client::module::{module_by_name, InnerDefinition};
 use crate::client::process::run_task;
 use anyhow::{anyhow, bail, Result};
 
-pub fn run_task_cmd(task_name: &str, _cli_config: &CliOptions) -> Result<()> {
-    let module_defs = read_module_definitions()?;
+pub fn run_task_cmd(task_name: &str, cli_config: &CliOptions) -> Result<()> {
+    let module_defs = read_module_definitions(&cli_config)?;
     let module_def =
         module_by_name(task_name, &module_defs).ok_or_else(|| {
             anyhow!("Failed to find task with name '{}'", task_name)